test: unify e2e suite with backend dispatch + run full suite against HAOS (refs #1281)
Single fixture chain, two backends. HAOS_TEST_IMAGE_PATH switches
ha_container_with_fresh_config from testcontainer to a booted qcow2;
ha_client and mcp_server pick up the right token from the dict.

Layout:
- New tests/src/haos_runtime.py: shared QEMU + login-flow helpers.
- tests/src/e2e/conftest.py: dispatch at the top of
  ha_container_with_fresh_config; token-aware ha_client / mcp_server;
  pytest_collection_modifyitems hook enforces haos_only / container_only
  markers (registered in pytest.ini).
- tests/src/e2e/haos_only/: tests that only make sense against HAOS
  (canary moved here from the deleted tests/src/haos_e2e/ package).
- Workflow now runs `pytest src/e2e/` with HAOS_TEST_IMAGE_PATH set.
  Same test code, different backend. Container-only tests skip
  automatically when the marker is present (none yet — triage in a
  follow-up commit on this PR).

This is the unified architecture the issue calls for: PR authors write
one test, both harnesses exercise it (unless one is physically
impossible, in which case they use the appropriate marker).

closes #1281
