+++ tests/pytest.ini
@@ -51,7 +51,8 @@ markers =
     core: core tool tests (state, service, bulk control, history, templates)
     config: configuration management tests (helpers, labels)
     filesystem: filesystem access tools tests
-    haos: tests that target the HAOS-QEMU tier (require pre-baked qcow2; #1281)
+    haos_only: skip unless HAOS_TEST_IMAGE_PATH is set (HAOS-QEMU backend required; #1281)
+    container_only: skip when HAOS backend is active (testcontainer-specific behavior; #1281)
 
 # Async support — session-scoped loop lets session-scoped fixtures
 # (mcp_server, mcp_client, ha_client) share one event loop with tests.
