--- tests/src/haos_e2e/__init__.py
@@ -1,10 +1,0 @@
-"""HAOS-tier E2E tests (see #1281).
-
-Tests in this package run against a pre-baked HAOS qcow2 booted under
-QEMU/KVM rather than the testcontainer Docker image used by the rest of
-the ``e2e/`` suite. They opt in via ``@pytest.mark.haos`` and are
-scheduled by ``.github/workflows/haos-e2e-tests.yml``.
-
-Sibling of ``e2e/`` rather than nested so the parent ``e2e/conftest.py``
-testcontainer session fixture isn't auto-triggered for haos runs.
-"""
